backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pymysql==1.1.1
pydantic-settings==2.10.1
email-validator==2.2.0
python-jose[cryptography]==3.5.0
